+++ reproc/test/working-directory.cpp
@@ -12,16 +12,8 @@ TEST_SUITE("reproc")
     REPROC_ERROR error = REPROC_SUCCESS;
     INFO(reproc_strerror(error));
 
-    // Executable path is relative to reproc/resources because we change
-    // working directory before executing noop as a child process.
     const char *working_directory = "reproc/resources";
-
-#if defined(_WIN32)
-    std::array<const char *, 2> argv{ "reproc/resources/noop",
-                                              nullptr };
-#else
-    std::array<const char *, 2> argv{ "./noop", nullptr };
-#endif
+    std::array<const char *, 2> argv{ "reproc/resources/noop", nullptr };
 
     error = reproc_start(&noop, argv.data(), working_directory);
     REQUIRE(!error);